Add Environment::platform() and Environment::path()
The platform() method detects the current platform and is comparable with the PHP
constant PHP_OS_FAMILY - however, it recognizes Cygwin and also works for PHP 7.0
and PHP 7.1, where this constant is not yet defined.

The path() method formats a path to a short format usable in the given platform,
e.g. using ".", ".." and "~" to reference the current, parent and home directories,
as well as $USERPOFILE and $APPDATA on Windows